requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.2.1
Django==3.0.5
django-widget-tweaks==1.4.8
importlib-metadata==1.6.0
include==0.2.2
kombu==4.6.8
phonenumbers==8.12.1
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
pycountry==18.12.8
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
